+++ test/matrix_list.jl
@@ -14,3 +14,15 @@ g2 = MatrixList(g)
 @test nodes(g2) == nodes(g)
 @test timestamps(g2) == timestamps(g)
 @test matrix(g2, 1) == matrix(g, "t1")
+
+g = MatrixList()
+@test isempty(g)
+add_matrix!(g, speye(4))
+@test isempty(g) == false
+I = [1,2,1]
+J = [1,2,4]
+V = ones(3)
+A = sparse(I,J,V, 4,4)
+add_matrix!(g, A)
+@test num_nodes(g) == 4
+@test num_timestamps(g) == 2